+++ src/pages/Counsel/index.tsx
@@ -1,6 +1,6 @@
 import Head from 'next/head';
 
-import { getTerms } from '@/src/api/terms';
+import { getTerms } from '@/src/utils/fetch/terms';
 import Footer from '@/src/components/Footer/Footer';
 import Header from '@/src/components/Header';
 import Layout from '@/src/components/Layout/Layout';
@@ -9,7 +9,7 @@ import CounselScreen from '@/src/screens/Counsel/CounselScreen';
 
 import type { GetServerSideProps } from 'next';
 import type { NextPageWithLayout } from '@/src/components/Layout/types';
-import type { Term } from '@/src/api/terms';
+import type { Term } from '@/src/utils/fetch/terms';
 
 interface Props {
   terms: Term[];
