+++ src/pages/faq/index.tsx
@@ -2,9 +2,9 @@ import type { GetServerSideProps, GetServerSidePropsContext } from 'next';
 import Head from 'next/head';
 import React from 'react';
 
-import { getFaqCategories } from '@/src/api/faq/getFaqCategories';
-import { getFaqs } from '@/src/api/faq/getFaqs';
-import { getTerms, Term } from '@/src/api/terms';
+import { getFaqCategories } from '@/src/utils/fetch/faq/getFaqCategories';
+import { getFaqs } from '@/src/utils/fetch/faq/getFaqs';
+import { getTerms, Term } from '@/src/utils/fetch/terms';
 import Footer from '@/src/components/Footer/Footer';
 import Header from '@/src/components/Header';
 import Layout from '@/src/components/Layout/Layout';
