+++ src/utils/fetch/faq/getFaqs.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 
-import { getApiUrl } from '@/src/utils/api';
+import { getApiUrl } from '@/src/utils/apiHelper';
 
 import { FaqListResponseSchema } from './schema';
